refactor: 스크립트 중복 유틸 통합과 죽은 코드 정리
- scripts/lib.js 신설: CRC32, manifest 읽기, getRequiredEnv,
  readResponseBody, getPackageFile을 한 곳으로 통합
- src/shared.js의 미사용 matchesSearch 삭제(스펙 API 표 함께 갱신)
- popup.html 인라인 헤더 스타일을 popup.css로 이동, 충돌하던
  죽은 header 규칙 제거
- options.css 미사용 .sr-only 삭제, 두 CSS의 중복 선언된
  .theme-status-visible 블록 병합
- version.js: 불필요한 async 제거와 따옴표 통일
- docs/landing.js: normalizeSearchText의 중복 trim 제거

동작 변경 없음. npm test 49개, validate:data/dark-icons/landing-data,
package, publish dry run 전부 통과 확인.

Close #134
