+++ src/popup.css
@@ -19,16 +19,28 @@ body {
 
 header {
   display: flex;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
-  margin-bottom: 15px;
+  margin-bottom: 10px;
+}
+
+.header-title {
+  display: flex;
+  align-items: baseline;
 }
 
 header h1 {
   margin: 0;
   font-size: 20px;
   font-weight: 700;
+  line-height: 1;
   color: var(--color-primary);
+}
+
+#current-version {
+  font-size: 12px;
+  color: var(--color-text-muted);
+  margin-left: 8px;
 }
 
 .search-bar {
@@ -334,11 +346,8 @@ header h1 {
   overflow: visible;
   clip: auto;
   white-space: normal;
-  color: var(--color-notice);
-  text-align: center;
-}
-
-.theme-status-visible {
   margin: 0 0 8px;
+  color: var(--color-notice);
   font-size: 11.5px;
+  text-align: center;
 }
