+++ src/popup.html
@@ -11,31 +11,10 @@
   </head>
   <body>
     <div class="container">
-      <header
-        style="
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          width: 100%;
-          margin-bottom: 10px;
-        "
-      >
-        <div style="display: flex; align-items: baseline;">
-          <h1
-            style="
-              margin: 0;
-              font-size: 20px;
-              font-weight: 700;
-              color: var(--color-primary);
-              line-height: 1;
-            "
-          >
-            LinKHU
-          </h1>
-          <span
-            id="current-version"
-            style="font-size: 12px; color: var(--color-text-muted); margin-left: 8px;"
-          ></span>
+      <header>
+        <div class="header-title">
+          <h1>LinKHU</h1>
+          <span id="current-version"></span>
           <span id="update-message"></span>
         </div>
