+++ tests/shared.test.js
@@ -9,18 +9,6 @@ test("shared normalize removes spaces and letter case", () => {
   assert.equal(LinKHUShared.normalize(null), "");
 });
 
-test("shared matchesSearch matches name, id, and category", () => {
-  const site = { id: "software", name: "소프트웨어 융합대학", category: "단과대" };
-
-  assert.equal(
-    LinKHUShared.matchesSearch(site, LinKHUShared.normalize("소프트웨어융합 대학")),
-    true,
-  );
-  assert.equal(LinKHUShared.matchesSearch(site, "software"), true);
-  assert.equal(LinKHUShared.matchesSearch(site, "단과대"), true);
-  assert.equal(LinKHUShared.matchesSearch(site, "없는검색어"), false);
-});
-
 test("shared rankSites orders by name prefix, name, id, then category", () => {
   const sites = [
     { id: "swcon", name: "소프트웨어융합학과", category: "학과" },
